Implement blacklisting
Implement the option to specify a list of projects that
should be ignored.

The idea is to create a deep copy of the user-specified
projects array, and then take the complement of this set
of projects iff blacklistProjects === true.

I opted to rebuild this complement list for every fetch,
because it is possible that the user creates a new project
on Todoist during the lifetime of the program.